fix: Resolve build errors and refactor API handlers
