+++ package.json
@@ -46,9 +46,9 @@
     "@supabase/ssr": "^0.6.1",
     "@supabase/supabase-js": "^2.39.8",
     "@types/cheerio": "^0.22.35",
-    "@types/node": "20.6.2",
-    "@types/react": "18.2.22",
-    "@types/react-dom": "18.2.7",
+    "@types/node": "^22.14.0",
+    "@types/react": "^19.1.0",
+    "@types/react-dom": "^19.1.2",
     "autoprefixer": "10.4.15",
     "cheerio": "^1.0.0",
     "class-variance-authority": "^0.7.0",
@@ -57,11 +57,11 @@
     "date-fns": "^3.6.0",
     "embla-carousel-react": "^8.3.0",
     "eslint": "8.49.0",
-    "eslint-config-next": "13.5.1",
+    "eslint-config-next": "^15.3.0",
     "googleapis": "^148.0.0",
     "input-otp": "^1.2.4",
     "lucide-react": "^0.446.0",
-    "next": "13.5.1",
+    "next": "^15.3.0",
     "next-auth": "^4.24.11",
     "next-themes": "^0.3.0",
     "pino": "^9.6.0",
@@ -77,7 +77,7 @@
     "tailwind-merge": "^2.5.2",
     "tailwindcss": "3.3.3",
     "tailwindcss-animate": "^1.0.7",
-    "typescript": "5.2.2",
+    "typescript": "^5.8.3",
     "vaul": "^0.9.9",
     "zod": "^3.23.8"
   }
